Factor out truncation hook.
